code-style: use a more appropriate name
unlike most programs where there's a single entry point and everything
else is called from within that master function - in mpv scripts the
functions get called whenever mpv decides to call them.

so this is less of a master function and more of an initialization
function that sets up the callback handlers.