mfpbar: revise mouse handling
instead of activating the bar when the mouse is in proximity, this
changes it so that he previous mouse coordinate had to be in proximity
as well.

in theory this introduces a bit of latency, but seems to be unavoidable
as otherwise the old mouse coordinate can end up incorrectly activating
the bar during window resize.

the patch also changes the proximity check so that mouse that's way to
far below won't activate the bar either.

ref: https://codeberg.org/NRK/mpv-toolbox/issues/4
fixes: https://codeberg.org/NRK/mpv-toolbox/issues/9