mfpbar: add chapter snapping support
kinda nice. but also creates some ambiguities, for example if
the user wanted to seek to near a chapter instead of snapping
onto it.

might need to add another binding such as shift+click or
ctrl+click for exact seek. or maybe chapter snapping shouldn't
be part of mfpbar to begin with since mdmenu can jump to a
chapter already.

for now, i'll just merge this. will deal with exact-seek issue
later and if needed.